General

+++ src/qc_hamiltonian.hpp
@@ -28,6 +28,6 @@ namespace py = pybind11;
 using namespace std;
 
 vector<tuple<py::array_t<uint32_t>, py::array_t<uint32_t>, py::array_t<double>,
-             py::array_t<uint32_t>>>
+             py::array_t<uint64_t>>>
 build_qc_mpo(py::array_t<int32_t> orb_sym, py::array_t<double> t,
              py::array_t<double> v);
